+++ src/processors/globalStyle.ts
@@ -1,0 +1,14 @@
+import { PreprocessorGroup } from '../typings';
+
+export default (): PreprocessorGroup => {
+  return {
+    async style({ content, attributes, filename }) {
+      const { default: transformer } = await import(
+        '../transformers/globalStyle'
+      );
+      if (!attributes.global) return { code: content };
+
+      return transformer({ content, filename });
+    },
+  };
+};