+++ src/processors/pug.js
@@ -1,7 +1,0 @@
-const transformer = require('../transformers/pug.js');
-
-module.exports = options => ({
-  markup({ content, filename }) {
-    return transformer({ content, filename, options });
-  },
-});