+++ test/transformers/stylus.test.ts
@@ -1,0 +1,15 @@
+import { resolve } from 'path';
+
+import getAutoPreprocess from '../../src';
+import { preprocess } from '../utils';
+
+describe('transformer - stylus', () => {
+  it('should return @imported files as dependencies', async () => {
+    const template = `<style lang="stylus">@import "fixtures/style.styl";</style>`;
+    const opts = getAutoPreprocess();
+    const preprocessed = await preprocess(template, opts);
+    expect(preprocessed.dependencies).toContain(
+      resolve(__dirname, '..', 'fixtures', 'style.styl'),
+    );
+  });
+});